Moving files from abandoned schema-2012-06 branch
All of the changes to specification folder in the
schema-2012-06 branch are being moved over to dev_4_4
and develop branches ahead of the restructuring of
the specification folder.

Conflicts:
	components/specification/Schemas/BinaryFile/2011-06/index.html
	components/specification/Schemas/OME/2011-06/index.html
	components/specification/Schemas/ROI/2011-06/index.html
	components/specification/Schemas/SA/2011-06/index.html
	components/specification/Schemas/SPW/2011-06/index.html
	components/specification/Schemas/index.html

These files were modified for the release of 2012 schema.
They hare been removed as they are no longer needed as generated
from template by website deploy script.